lesson 1.2 react dom and components
